+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # ChangeLog
 
-* add method to modify recipelogic runtime data
-* fix kjs events (onRecipeModify => onBeforeRecipeModify + onAfterRecipeModify)
 * rework the `ICapabilityProviderTrait` to allow a trait to provide multiple capabilities
 * add spark attachable to the mana trait
+* add redstone signal condition